fix: styles components
