+++ src/pages/catalog-product/CatalogProduct.tsx
@@ -1,5 +1,5 @@
 import { ReactElement } from 'react';
-import BasicTabs from './card-car/ShowProducts';
+import BasicTabs from './show-cars/ShowCars';
 
 export default function CatalogProduct(): ReactElement {
   return <BasicTabs />;
