+++ src/pages/catalog-product/card-car/card-car.module.css
@@ -44,7 +44,6 @@
 .old-price {
   margin: 0;
   text-align: right;
-  padding-right: 1rem;
   font-size: large;
 }
 
@@ -67,7 +66,17 @@
 }
 
 @media (max-width: 600px) {
-  .card-title {
+  .card-title,
+  .card-description {
+    font-size: 1rem;
+  }
+
+  .card-description {
+    font-size: 0.8rem;
+  }
+
+  .discount-price,
+  .old-price {
     font-size: 1rem;
   }
 }
